Prepare for publish (2.0.0)
