+++ Cargo.toml
@@ -2,7 +2,7 @@
 authors = ["Connor Slade <connor@connorcode.com>"]
 edition = "2018"
 name = "afire"
-version = "2.0.0-alpha"
+version = "2.0.0"
 
 categories = ["network-programming", "web-programming::http-server"]
 description = "🔥 A blazing fast web framework for Rust"
