+++ lib/extensions/mod.rs
@@ -3,4 +3,3 @@ pub mod logger;
 pub mod ratelimit;
 pub mod request_id;
 pub mod serve_static;
-pub mod util;
